pyproject.toml
[project]
name = "jarvis"
version = "0.1.0"
description = "개인 AI 비서 시스템 - Google Calendar + Gmail MCP 서버"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
jarvis = "jarvis.server:main"
jarvis-auth = "jarvis.auth.google_auth:authenticate"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jarvis"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
